Fix/rm kommuner from url (#168)
* rm /kommuner/ path + create 404 pag

* rm folder kommuner
