+++ components/Layout.tsx
@@ -129,9 +129,9 @@ export default function Layout({ children }: { children: JSX.Element }) {
         <Nav path={path}>
           <ul>
             <li>
-              <Link href="/kommuner">
-                <a className={path == "/kommuner" ? "active" : "inactive"}
-                href="/kommuner">
+              <Link href="/">
+                <a className={path == "/" ? "active" : "inactive"}
+                href="/">
                   Kommuner
                 </a>
               </Link>
